requirements.txt
# Carbon Sentinel — Python Dependencies
# Install with:  pip install -r requirements.txt

# ── Web / API ─────────────────────────────────────────────────────────────────
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0

# ── Blockchain / Web3 ─────────────────────────────────────────────────────────
web3>=6.15.0

# ── Supabase ─────────────────────────────────────────────────────────────────
supabase>=2.4.0

# ── Environment ───────────────────────────────────────────────────────────────
python-dotenv>=1.0.0

# ── Typing / Pydantic ────────────────────────────────────────────────────────
# (used implicitly by supabase-py)
httpx>=0.26.0
